pyproject.toml
[project]
name = "langchain-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipython>=9.5.0",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.10",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
]
